[grid] Restructuring classes have stateful data in LocalDistributor
